feat (tournament): add a bunch of admin functions
See also:
- WAK-31 Referee can check present players of team
- WAK-32 Referee can register death order

+++ back/src/main/java/com/waktoolbox/waktool/domain/models/tournaments/TeamStatsByClass.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TeamStatsByClass {
-    Integer id;
+    Byte id;
     Integer death;
     Integer banned;
     Integer killed;
